Added a vitals buffer to the database helper.
Buffered data is stored in a map of integer pids to lists of vitals objects. This data structure is stored within a private inner class of the db helper. A static instance of the buffer is stored in the db helper.

The buffer inserts vitals in a sliding window scheme with a default window size of 100.

All vital-related messages observed by the DatabaseMessageListener are buffered.

The update method in the DatabaseMessageListener has been slimmed down by replacing most of the temporary entries list with instances of the appropriate models (Patient & Vital)

The models classes have been modified to support null attribute values in the toEntriesListMethod